+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorIntegerTest.kt
@@ -1,0 +1,125 @@
+/*
+ * @(#) CodeGeneratorIntegerTest.kt
+ *
+ * json-kotlin-schema-codegen  JSON Schema Code Generation
+ * Copyright (c) 2021 Peter Wall
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package net.pwall.json.schema.codegen
+
+import kotlin.test.Test
+import kotlin.test.expect
+
+import java.io.File
+import java.io.StringWriter
+
+import net.pwall.json.schema.codegen.CodeGeneratorTestUtil.createHeader
+import net.pwall.json.schema.codegen.CodeGeneratorTestUtil.outputCapture
+
+class CodeGeneratorIntegerTest {
+
+    @Test fun `should generate correct code for integer formats`() {
+        val input = File("src/test/resources/test-integer-format.schema.json")
+        val codeGenerator = CodeGenerator()
+        codeGenerator.baseDirectoryName = "dummy"
+        val stringWriter = StringWriter()
+        codeGenerator.outputResolver = outputCapture("dummy", emptyList(), "TestIntegerFormat", "kt", stringWriter)
+        codeGenerator.basePackageName = "com.example"
+        codeGenerator.generate(input)
+        expect(createHeader("TestIntegerFormat") + expected) { stringWriter.toString() }
+    }
+
+    @Test fun `should generate correct code for integer formats in Java`() {
+        val input = File("src/test/resources/test-integer-format.schema.json")
+        val codeGenerator = CodeGenerator(templates = "java", suffix = "java")
+        codeGenerator.baseDirectoryName = "dummy"
+        val stringWriter = StringWriter()
+        codeGenerator.outputResolver = outputCapture("dummy", emptyList(), "TestIntegerFormat", "java", stringWriter)
+        codeGenerator.basePackageName = "com.example"
+        codeGenerator.generate(input)
+        expect(createHeader("TestIntegerFormat") + expectedJava) { stringWriter.toString() }
+    }
+
+    companion object {
+
+        const val expected =
+"""package com.example
+
+/**
+ * Test integer formats.
+ */
+data class TestIntegerFormat(
+    val little: Int,
+    val big: Long
+)
+"""
+
+        const val expectedJava =
+"""package com.example;
+
+/**
+ * Test integer formats.
+ */
+public class TestIntegerFormat {
+
+    private final int little;
+    private final long big;
+
+    public TestIntegerFormat(
+            int little,
+            long big
+    ) {
+        this.little = little;
+        this.big = big;
+    }
+
+    public int getLittle() {
+        return little;
+    }
+
+    public long getBig() {
+        return big;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (this == other)
+            return true;
+        if (!(other instanceof TestIntegerFormat))
+            return false;
+        TestIntegerFormat typedOther = (TestIntegerFormat)other;
+        if (little != typedOther.little)
+            return false;
+        return big == typedOther.big;
+    }
+
+    @Override
+    public int hashCode() {
+        int hash = little;
+        return hash ^ (int)big;
+    }
+
+}
+"""
+
+    }
+
+}